azure-func-python-terraform/pyproject.toml
[project]
name = "azure-func-python-ref"
version = "0.1.0"
description = "Reference Azure Functions (Python) + Terraform project"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.20.0",
    "pydantic>=2.6",
    "structlog>=24.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
